requirements.txt
fastapi
pydantic
sqlalchemy
uvicorn
psycopg2-binary
python-dotenv
pyjwt
pytest
pytest-cov
httpx
requests